Change ConfigLoader to be a store of ConfigDelta not a store of ConfigValues.
The main function of the loader is to process the override hierarchy of classpath < api < overrides.  api responses can now be more complex with defaults, vs env specific. But we don't care about that. Make Overrides/Classpath act like the API ones so we get our basic clobbering behavior back.