create the concept of default_env
When running tests locally, I may want different defaults than normal. I could use a regular Prefab environment, but what if I want to be in local_mode. In that case, let me specify a section in the default yaml for test values and choose to use that with 'default_env' in options.

These envs are != prefab project environments. That's intentional. Users shouldn't have to know the ID of their projectEnv.