Add EvaluatedKeysAggregator (disabled by default)
The EvaluatedKeysAggregator phones home with the config keys that have
been evaluated.

This allows us to identify zombie config/flags/etc. that are no longer
in use as well as to generate usage for active config/flags/etc.

You can opt-in by specifying `collect_keys: true` in the client
initialization options.