fix an issue similar to https://github.com/crohr/syslogger/issues/29 (#164)
With #161 we added support for tagged logging. Unfortunately Rails assumes that if you support `tagged` you must be a TaggedLogger and then proceeds to call the method `current_tags` on your internal formatter object, which prevents it from being a simple proc.

This is the minimally viable PR to retain simple formatter procs, while also allowing ActiveJob to inspect our internals.