remove the assert_no_logging
The fact is, we do log from much of this code. Whether or not it ends up on stdout is a matter of config and whether or not the log_filter has been turned on. Better to test that directly in test_internal_logger. It was very tough to get working here when tests parallelize.